Library Management Project Files Create Successfully

+++ Library_Management_System/src/main/java/com/library/Library_Management_System/Services/BorrowService.java
@@ -5,8 +5,8 @@ import com.library.Library_Management_System.Entity.Borrow;
 import java.util.List;
 
 public interface BorrowService {
-    Borrow borrowSaver(Long userId, Long bookId);
+    Borrow borrowSaver(int userId, int bookId);
     List<Borrow> getAllDetails();
-    Borrow getById(Long id);
-    void deleteById(Long id);
+    Borrow getById(int id);
+    void deleteById(int id);
 }
